Ibero_Microcredenciales_Entregable/microcredentials_app/requirements.txt
streamlit>=1.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
pdfplumber>=0.9.0
requests>=2.28.0
beautifulsoup4>=4.12.0
